dedicated jobs for event schedule update
