+++ Sources/swiftarr/Resources/Views/admin/scheduleUpload.html
@@ -23,13 +23,35 @@
 							<div class="row mb-2"> 
 								<input type="file" accept=".ics" name="schedule" required>
 							</div>
-							<div class="alert alert-danger mt-3 d-none" role="alert">
-							</div>		
+							<div class="alert alert-danger mt-3 d-none" role="alert"></div>
 							<div class="row mb-2">
 								<div class="col col-auto gx-0">
 									<button type="submit" class="btn btn-primary">Upload<span class="spinner-border spinner-border-sm d-none" role="status" aria-hidden="true"></span></button>
 								</div>
 							</div>
+						</div>
+					</form>
+				</li>
+			</ul>
+			<div class="row mb-2">
+				<div class="col">
+					<b>Manual Schedule Update: Run schedule update job</b>
+				</div>
+			</div>
+			<ul class="container-md mx-0 px-0 list-group">
+				<li class="list-group-item bg-transparent mb-3">				
+					<form class="ajax" action="/admin/schedulereload" enctype="multipart/form-data" method="POST" data-successurl="reset" id="schedulereloadform">
+						<div class="container-fluid">
+							<div class="alert alert-danger mt-3 d-none" role="alert">
+							</div>		
+							<div class="row mb-2">
+								<div class="col col-auto gx-0">
+									<button type="submit" class="btn btn-primary">Reload<span class="spinner-border spinner-border-sm d-none" role="status" aria-hidden="true"></span></button>
+								</div>
+							</div>
+						</div>
+						<div class="alert alert-success mt-3 d-none" role="alert">
+							Job submitted! Refresh this page in a minute or so to view the result.
 						</div>
 					</form>
 				</li>
